Move LocationBroadcastReceiver.kt into sensors package.
