+++ app/src/main/AndroidManifest.xml
@@ -77,7 +77,7 @@
         </activity>
 
         <receiver
-            android:name=".background.LocationBroadcastReceiver"
+            android:name=".sensors.LocationBroadcastReceiver"
             android:enabled="true"
             android:exported="true">
             <intent-filter>
