--- app/src/main/java/io/homeassistant/companion/android/background/ReceiverComponent.kt
@@ -1,10 +1,0 @@
-package io.homeassistant.companion.android.background
-
-import dagger.Component
-import io.homeassistant.companion.android.common.dagger.AppComponent
-
-@Component(dependencies = [AppComponent::class])
-interface ReceiverComponent {
-
-    fun inject(receiver: LocationBroadcastReceiverBase)
-}
