+++ app/src/main/java/io/homeassistant/companion/android/sensors/SensorComponent.kt
@@ -7,6 +7,8 @@ import io.homeassistant.companion.android.common.dagger.AppComponent
 @Component(dependencies = [AppComponent::class], modules = [PresenterModule::class])
 interface SensorComponent {
 
+    fun inject(locationBroadcastReceiver: LocationBroadcastReceiver)
+
     fun inject(worker: SensorWorker)
 
     fun inject(sensorsSettingsFragment: SensorsSettingsFragment)
