+++ app/src/main/java/io/homeassistant/companion/android/util/PermissionManager.kt
@@ -7,8 +7,7 @@ import android.content.pm.PackageManager
 import android.os.Build
 import androidx.core.app.ActivityCompat
 import androidx.fragment.app.Fragment
-import io.homeassistant.companion.android.background.LocationBroadcastReceiver
-import io.homeassistant.companion.android.background.LocationBroadcastReceiverBase
+import io.homeassistant.companion.android.sensors.LocationBroadcastReceiver
 
 class PermissionManager {
 
@@ -66,7 +65,7 @@ class PermissionManager {
 
         fun restartLocationTracking(context: Context) {
             val intent = Intent(context, LocationBroadcastReceiver::class.java)
-            intent.action = LocationBroadcastReceiverBase.ACTION_REQUEST_LOCATION_UPDATES
+            intent.action = LocationBroadcastReceiver.ACTION_REQUEST_LOCATION_UPDATES
 
             context.sendBroadcast(intent)
         }
