+++ app/src/minimal/java/io/homeassistant/companion/android/sensors/LocationBroadcastReceiver.kt
@@ -1,0 +1,29 @@
+package io.homeassistant.companion.android.sensors
+
+import android.content.BroadcastReceiver
+import android.content.Context
+import android.content.Intent
+
+class LocationBroadcastReceiver : BroadcastReceiver() {
+
+    companion object {
+        const val MINIMUM_ACCURACY = 200
+
+        const val ACTION_REQUEST_LOCATION_UPDATES =
+            "io.homeassistant.companion.android.background.REQUEST_UPDATES"
+        const val ACTION_REQUEST_ACCURATE_LOCATION_UPDATE =
+            "io.homeassistant.companion.android.background.REQUEST_ACCURATE_UPDATE"
+        const val ACTION_PROCESS_LOCATION =
+            "io.homeassistant.companion.android.background.PROCESS_UPDATES"
+        const val ACTION_PROCESS_GEO =
+            "io.homeassistant.companion.android.background.PROCESS_GEOFENCE"
+
+        internal const val TAG = "LocBroadcastReceiver"
+    }
+
+    override fun onReceive(context: Context, intent: Intent) {
+        //Noop
+    }
+
+
+}
